openapi: Combine both openapi markdown extension to a single extension.
Currently api_description and api_code_examples are the two markdown
extensions related to openapi. Combine them to a single markdown
extension created in the openapi folder for clearer understanding
of code and also for a better development of the future markdown
extensions related to openapi.
